Feature/updates (#4)
* Improved messaging in some tests AND added test of multiple levels of decrementation

* Renamed DEFAULT_LEVELS to DEFAULT_SPACES_PER_LEVEL

* Implemented unit test decrementMultipleLevels

* Restructured Project folder

* Removed unused dependencies and updated build.gradle

* Refactored decrement method

* Refactored format method

* Updated license file and license header in the classes

* Added Jacoco for coverage

* Updated UML diagram class

* Updated name of field in class diagram

* Fixed PR comments
