+++ build.gradle
@@ -1,0 +1,34 @@
+plugins {
+    id 'java'
+    id 'jacoco'
+}
+
+version = '1.0.7'
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testImplementation 'org.testng:testng:7.5'
+}
+
+test {
+    useTestNG()
+    finalizedBy jacocoTestReport
+}
+
+wrapper {
+    gradleVersion = 7.3
+}
+
+jacoco {
+    toolVersion = "0.8.7"
+}
+
+jacocoTestReport {
+    reports {
+        xml.required = false
+        csv.required = false
+        html.outputLocation = layout.buildDirectory.dir('reports/jacocoReports')
+    }
+}
