--- indentmanager/build.gradle
@@ -1,26 +1,0 @@
-
-plugins {
-    id 'java-library'
-}
-
-version = '1.0.7'
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation 'org.testng:testng:7.4.0'
-    api 'org.apache.commons:commons-math3:3.6.1'
-    implementation 'com.google.guava:guava:30.1-jre'
-}
-
-tasks.named('test') {
-    useTestNG()
-}
-
-tasks.named('jar') {
-    manifest {
-        attributes('Implementation-Title': project.name,
-                'Implementation-Version': project.version)
-    }
-}
