+++ src/test/java/com/softwareonpurpose/indentmanager/IndentManagerTest.java
@@ -33,8 +33,15 @@ public class IndentManagerTest {
     }
 
     @DataProvider
-    public static Object[][] decrementingScenarios() {
+    public static Object[][] decrementOnceScenarios() {
         return new Object[][]{{1, ""}, {2, "  "}, {3, "    "}, {0, ""}};
+    }
+
+    @DataProvider
+    public static Object[][] decrementLevelsScenarios() {
+        return new Object[][]{
+                {0},{5},{9}
+        };
     }
 
     @Test
@@ -45,7 +52,7 @@ public class IndentManagerTest {
 
     @Test
     public void indentationOnDefaultInstantiation() {
-        String message = "On default instantiation, indent failed to be two spaces";
+        String message = "On default instantiation, indent failed to be zero spaces";
         Assert.assertEquals(IndentManager.getInstance().format("."), ".", message);
     }
 
@@ -133,7 +140,7 @@ public class IndentManagerTest {
         Assert.assertEquals(manager.format(message), String.format("%s%s", indentation, message), failureMessage);
     }
 
-    @Test(dataProvider = "decrementingScenarios")
+    @Test(dataProvider = "decrementOnceScenarios")
     public void decrementOnce(int levelsToIncrement, String expectedIndentation) {
         final String message = "Message Text";
         final IndentManager manager = IndentManager.getInstance();
@@ -143,5 +150,18 @@ public class IndentManagerTest {
         String failureMessage = String.format(failureMessageFormat, levelsToIncrement, expectedIndentation);
         Assert.assertEquals(manager.format(message), String.format("%s%s", expectedIndentation, message),
                 failureMessage);
+    }
+
+    @Test(dataProvider = "decrementLevelsScenarios")
+    public void decrementMultipleLevels(int levelsToDecrement){
+        final String message = "Message Text";
+        String failureMessageFormat = "After incrementing by %d and decrementing by %d, indentation failed to be '%s'";
+        int levelsToIncrement = levelsToDecrement + 2;
+        String expectedIndentation = "    ";
+        String failureMessage = String.format(failureMessageFormat, levelsToIncrement, levelsToDecrement, expectedIndentation);
+        final IndentManager manager = IndentManager.getInstance();
+        manager.increment(levelsToIncrement);
+        manager.decrement(levelsToDecrement);
+        Assert.assertEquals(manager.format(message), String.format("%s%s", expectedIndentation, message), failureMessage);
     }
 }
